+++ R/FourPHFfit.bulk.R
@@ -101,8 +101,9 @@ FourPHFfit.bulk <- function(data, total.seeds.col, counts.intervals.cols,
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
-    stop('"intervals" are not uniform.')
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
+    warning("'intervals' are not uniform.")
   }
 
   # Check if total.seeds.col is of type numeric