+++ R/GermSpeed.R
@@ -174,7 +174,8 @@ GermSpeed <- function(germ.counts, intervals, partial = TRUE,
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }
 
@@ -227,7 +228,8 @@ GermSpeedAccumulated <- function(germ.counts, intervals, partial = TRUE,
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }