+++ R/GermSynchrony.R
@@ -118,7 +118,8 @@ GermSynchrony <- function(germ.counts, intervals, partial = TRUE) {
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }
 
@@ -157,7 +158,8 @@ GermUncertainty <- function(germ.counts, intervals, partial = TRUE) {
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }