+++ R/MeanGermPercent.R
@@ -138,7 +138,8 @@ MeanGermPercent <- function(germinated.seeds, germ.counts, total.seeds, interval
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }
 
@@ -162,7 +163,8 @@ MeanGermNumber <- function(germ.counts, intervals, partial = TRUE) {
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }