+++ R/WeightGermPercent.R
@@ -84,7 +84,8 @@ WeightGermPercent <- function(germ.counts, total.seeds, intervals,
   }
 
   # Check if intervals are uniform
-  if (length(unique(diff(intervals))) != 1) {
+  idiff <- diff(intervals)
+  if (!all(abs(idiff - idiff[[1]]) < .Machine$double.eps ^ 0.5)) {
     warning("'intervals' are not uniform.")
   }